+++ src/main/java/org/tomas/testmlib/component/mutant/MutantComponentImpl.java
@@ -77,7 +77,7 @@ class MutantComponentImpl implements MutantComponent {
 
     long humans = dnas.size() - mutants;
 
-    double ratio = mutants / humans;
+    float ratio = (float) mutants / (float) humans;
     return new MutantStat(mutants, humans, ratio);
 
   }
